Move evaluate to mathWorker.js to make merging upstream easier
